+++ internal/helper/repo_test.go
@@ -77,6 +77,12 @@ func TestGetRepoPath(t *testing.T) {
 			err:      codes.InvalidArgument,
 		},
 		{
+			desc:     "storage defined but storage dir does not exist",
+			storages: []config.Storage{{Name: "default", Path: "/does/not/exist"}},
+			repo:     &pb.Repository{StorageName: "default", RelativePath: "foobar.git"},
+			err:      codes.Internal,
+		},
+		{
 			desc:     "relative path with directory traversal",
 			storages: exampleStorages,
 			repo:     &pb.Repository{StorageName: "default", RelativePath: "../bazqux.git"},