+++ CHANGELOG.md
@@ -8,6 +8,8 @@ UNRELEASED
   https://gitlab.com/gitlab-org/gitaly/merge_requests/257
 - Implement LastCommitForPath RPC
   https://gitlab.com/gitlab-org/gitaly/merge_requests/260
+- Deprecate Exists RPC in favor of RepositoryExists
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/260
 
 v0.26.0