+++ CHANGELOG.md
@@ -17,6 +17,8 @@ v0.27.0
   https://gitlab.com/gitlab-org/gitaly/merge_requests/260
 - Deprecate Exists RPC in favor of RepositoryExists
   https://gitlab.com/gitlab-org/gitaly/merge_requests/260
+- Install gems into vendor/bundle
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/264
 
 v0.26.0