+++ CHANGELOG.md
@@ -4,6 +4,8 @@ UNRELEASED
 
 - Increase gitaly-ruby connection timeout to 20s
   https://gitlab.com/gitlab-org/gitaly/merge_requests/265
+- Implement RepositorySize RPC
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/262
 
 v0.27.0